refactor(AppointmentOptionCommandService) : 인터페이스 분리 및 구현체 생성
- AppointmentOptionCommandService를 인터페이스로 분리하여 추상화 수준 향상
- 기존 서비스 클래스를 AppointmentOptionCommandServiceImpl로 구현
- 서비스 레이어의 구조를 개선하여 확장성과 테스트 용이성 향상
